Replace the license by an MIT one.
This can’t apply to the "code" in the .devcontainer and .github directories but there is almost nothing there, so I’ll assume that this is OK (these should remain under the original GPL license.

However, for the code for my layout, given that it is the QMK firmware that is linked with my layout, rather than the opposite, I believe that there is no issue to change the license.